feat(rgb): fill data to the digest
