+++ src/components/TransactionItem/TransactionRGBPPDigestModal/TransactionRGBPPDigestContent.tsx
@@ -1,15 +1,17 @@
 import { useTranslation } from 'react-i18next'
+import { Tooltip } from 'antd'
 import styles from './styles.module.scss'
 import { ReactComponent as CopyIcon } from '../../../assets/copy_icon.svg'
-import { Transaction, TransactionLeapDirection } from './types'
+import { TransactionLeapDirection } from './types'
 import { TransactionRGBPPDigestTransfer } from './TransactionRGBPPDigestTransfer'
 import { useSetToast } from '../../Toast'
+import { RGBDigest } from '../../../services/ExplorerService'
 
 export const TransactionRGBPPDigestContent = ({
   tx,
   leapDirection,
 }: {
-  tx: Transaction
+  tx: RGBDigest
   leapDirection: TransactionLeapDirection
 }) => {
   const { t } = useTranslation()
@@ -22,7 +24,9 @@ export const TransactionRGBPPDigestContent = ({
           <span>{t('address.seal_tx_on_bitcoin')}</span>
           <span className={styles.transactionHash}>{tx.txId}</span>
           <span className={styles.blockConfirm}>({tx.confirmations} Bitcoin Confirmed)</span>
-          <span className={styles.leap}>{t(`address.leap_${leapDirection}`)}</span>
+          <Tooltip placement="top" title={t(`address.leap_${leapDirection}_tip`)}>
+            <span className={styles.leap}>{t(`address.leap_${leapDirection}`)}</span>
+          </Tooltip>
         </div>
         <div className={styles.right}>
           <span>Commitment:</span>
