+++ src/components/TransactionItem/TransactionRGBPPDigestModal/styles.module.scss
@@ -94,6 +94,10 @@
   line-height: 14px;
 }
 
+.leap:hover {
+  cursor: pointer;
+}
+
 .left {
   display: flex;
   justify-content: flex-start;
