+++ src/locales/zh.json
@@ -384,7 +384,9 @@
       "transaction_rgbpp_digest": "RGB++ Digest",
       "seal_tx_on_bitcoin": "Seal Tx On Bitcoin",
       "leap_in": "Leap In",
-      "leap_out": "Leap Out"
+      "leap_out": "Leap Out",
+      "leap_in_tip": "本笔RGB++交易被标记为“Leap In”，即存在一般CKB资产转化为RGB++ 资产。",
+      "leap_out_tip": "本笔RGB++交易被标记为“Leap In”，即存在RGB++资产转化为一般CKB资产。"
     },
     "transaction": {
       "transaction": "交易",
