+++ src/services/ExplorerService/types.ts
@@ -283,11 +283,5 @@ export interface RGBDigest {
   txId: string
   confirmations: number
   commitment: string
-  transfers: {
-    address: string
-    transfers: {
-      capacity: string
-      cellType: string
-    }[]
-  }[]
+  transfers: TransactionRecord[]
 }
